# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api com autenticação pow JWT",
        "version": "0.1"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Email ou senha inválida"
                    }
                }
            }
        },
        "/api/teste": {
            "get": {
                "operationId": "App\\Http\\Controllers\\teste\\TesteController::index",
                "responses": {
                    "200": {
                        "description": "Lista de teste obtida com sucesso."
                    },
                    "400": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/api/teste/cadastrar": {
            "post": {
                "tags": [
                    "cadastrar teste"
                ],
                "operationId": "App\\Http\\Controllers\\teste\\TesteController::cadastrar",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Nome do teste",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "idade",
                        "in": "query",
                        "description": "Idade do teste",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Teste cadastrado com sucesso."
                    },
                    "400": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/api/teste/excluir": {
            "delete": {
                "operationId": "App\\Http\\Controllers\\teste\\TesteController::excluir",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id teste"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Teste removído com sucesso."
                    },
                    "400": {
                        "description": "Não autenticado"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearer": []
        }
    ]
}